+++ autoarray/structures/mock/mock_grid.py
@@ -1,14 +1,19 @@
 import numpy as np
 from typing import Tuple
 
+from autoarray.geometry.abstract_2d import AbstractGeometry2D
 from autoarray.inversion.linear_obj.neighbors import Neighbors
 from autoarray.structures.mesh.abstract_2d import Abstract2DMesh
 
 
-class MockGeometry:
+class MockGeometry(AbstractGeometry2D):
     def __init__(self, extent):
 
-        self.extent = extent
+        self._extent = extent
+
+    @property
+    def extent(self) -> np.ndarray:
+        return self._extent
 
 
 class MockGrid2DMesh(Abstract2DMesh):
